refactor(core): 将Hook模块加载功能提取到核心模块
- 将_load_hook_module_from_path和相关工具函数移动到app/core/hook.py
- 在AutoProxy.py中移除重复的Hook加载实现代码
