+++ app/core/__init__.py
@@ -27,7 +27,12 @@ __license__ = "GPL-3.0 license"
 from .broadcast import Broadcast
 from .config import Config, MaaConfig, GeneralConfig, MaaUserConfig, GeneralUserConfig
 from .emulator_manager import EmulatorManager
-from .hook import hookable, HookScope, HookPoint
+from .hook import (
+    hookable,
+    HookScope,
+    HookPoint,
+    load_hook_module,
+)
 from .task_manager import TaskManager
 
 from .timer import MainTimer
@@ -42,6 +47,7 @@ __all__ = [
     "HookPoint",
     "HookScope",
     "hookable",
+    "load_hook_module",
     "MainTimer",
     "TaskManager",
     "EmulatorManager",
