+++ app/core/hook.py
@@ -88,11 +88,24 @@ Hook 类型与语义
 from __future__ import annotations
 
 import contextvars
+import importlib.util
+import inspect
+import sys
 from collections.abc import Awaitable, Callable, Iterable
 from functools import wraps
+from hashlib import sha1
+from pathlib import Path
+from types import ModuleType
 from typing import Any, Concatenate, Literal, ParamSpec, TypeVar
 
-__all__ = ["hookable", "HookScope", "HookPoint"]
+from app.utils import get_logger
+
+__all__ = [
+    "hookable",
+    "HookScope",
+    "HookPoint",
+    "load_hook_module",
+]
 
 P = ParamSpec("P")
 R = TypeVar("R")
@@ -104,6 +117,146 @@ _ALL_HOOK_KINDS: frozenset[HookKind] = frozenset({"before", "after", "error", "a
 _current_scope: contextvars.ContextVar[HookScope | None] = contextvars.ContextVar(
     "hook_scope", default=None
 )
+
+logger = get_logger("Hook管理器")
+
+
+def _build_hook_logger(module: ModuleType, raw_path: str):
+    """为 Hook 模块构建专属 logger（优先使用 HOOK_META.name）。"""
+    hook_name: str | None = None
+    meta = getattr(module, "HOOK_META", None)
+    if isinstance(meta, dict):
+        name = meta.get("name")
+        if isinstance(name, str) and name.strip():
+            hook_name = name.strip()
+
+    if not hook_name:
+        stem = Path(raw_path).stem.strip()
+        hook_name = stem or getattr(module, "__name__", "unknown_hook")
+
+    return get_logger(f"hook-{hook_name}")
+
+
+def _inject_hook_logger(module: ModuleType, raw_path: str):
+    """为 Hook 模块创建并注入 logger。
+
+    - 注入 module.HOOK_LOGGER
+    - 若模块提供 set_logger(logger)，则一并调用
+    """
+    hook_logger = _build_hook_logger(module, raw_path)
+    setattr(module, "HOOK_LOGGER", hook_logger)
+
+    set_logger = getattr(module, "set_logger", None)
+    if callable(set_logger):
+        set_logger(hook_logger)
+
+    return hook_logger
+
+
+def _load_hook_module_from_path(file_path: str) -> tuple[ModuleType | None, str | None]:
+    """从 .py 文件路径动态加载模块。
+
+    返回 (module, warning)。失败时 module 为 None，warning 为说明文本。
+    """
+    p = Path(file_path)
+    if not p.exists() or not p.is_file():
+        return None, "文件不存在或不是文件"
+    if p.suffix.lower() != ".py":
+        return None, "仅支持 .py 文件"
+
+    try:
+        resolved = p.resolve()
+    except Exception:
+        resolved = p
+
+    # 生成稳定且尽量不冲突的模块名（避免重复 import 相互覆盖）
+    digest = sha1(str(resolved).encode("utf-8", errors="ignore")).hexdigest()[:12]
+    module_name = f"auto_mas_hook_{p.stem}_{digest}"
+
+    # 每次任务运行都重新加载一份，避免被上次运行的模块状态污染。
+    if module_name in sys.modules:
+        sys.modules.pop(module_name, None)
+
+    spec = importlib.util.spec_from_file_location(module_name, str(resolved))
+    if spec is None or spec.loader is None:
+        return None, "无法创建模块加载器"
+
+    module = importlib.util.module_from_spec(spec)
+    try:
+        spec.loader.exec_module(module)
+    except Exception as e:
+        return None, f"导入失败: {type(e).__name__}: {e}"
+
+    return module, None
+
+
+def load_hook_module(
+    hook_paths: list[str], scope: HookScope, target_cls: type[Any]
+) -> list[str]:
+    """按顺序加载并注册 Hook 列表。
+
+    - 每个 Hook 模块会注入 ``HOOK_LOGGER``（命名为 ``hook-xxx``）
+    - 兼容 register(scope, target_cls) / register(scope, target_cls, hook_logger)
+    - 发生错误时继续处理后续 Hook，并返回 warning 列表
+    """
+    warnings: list[str] = []
+
+    for raw_path in hook_paths:
+        module, warn = _load_hook_module_from_path(raw_path)
+        if warn:
+            msg = f"Hook 加载警告 [{raw_path}]: {warn}"
+            warnings.append(msg)
+            logger.warning(msg)
+            continue
+
+        assert module is not None
+        try:
+            hook_logger = _inject_hook_logger(module, raw_path)
+        except Exception as e:
+            msg = f"Hook 日志注入警告 [{raw_path}]: {type(e).__name__}: {e}"
+            warnings.append(msg)
+            logger.warning(msg)
+            continue
+
+        register = getattr(module, "register", None)
+        if not callable(register):
+            msg = f"Hook 加载警告 [{raw_path}]: 未找到可调用的 register(scope, target_cls)"
+            warnings.append(msg)
+            logger.warning(msg)
+            continue
+
+        try:
+            # 兼容两种签名：
+            # - register(scope, target_cls)
+            # - register(scope, target_cls, hook_logger)
+            sig = inspect.signature(register)
+            params = list(sig.parameters.values())
+            positional_count = len(
+                [
+                    p
+                    for p in params
+                    if p.kind
+                    in (
+                        inspect.Parameter.POSITIONAL_ONLY,
+                        inspect.Parameter.POSITIONAL_OR_KEYWORD,
+                    )
+                ]
+            )
+            has_varargs = any(
+                p.kind == inspect.Parameter.VAR_POSITIONAL for p in params
+            )
+
+            if has_varargs or positional_count >= 3:
+                register(scope, target_cls, hook_logger)
+            else:
+                register(scope, target_cls)
+        except Exception as e:
+            msg = f"Hook 注册警告 [{raw_path}]: {type(e).__name__}: {e}"
+            warnings.append(msg)
+            logger.warning(msg)
+            continue
+
+    return warnings
 
 
 # ---------------------------------------------------------------------------
