+++ app/task/general/AutoProxy.py
@@ -24,16 +24,16 @@ import uuid
 import shlex
 import shutil
 import asyncio
-import importlib.util
-import inspect
-import sys
-from types import ModuleType
-from hashlib import sha1
 from pathlib import Path
 from contextlib import suppress
 from datetime import datetime, timedelta
 
-from app.core import Config, HookScope, hookable
+from app.core import (
+    Config,
+    HookScope,
+    hookable,
+    load_hook_module,
+)
 from app.models.task import TaskExecuteBase, ScriptItem, LogRecord
 from app.models.ConfigBase import MultipleConfig
 from app.models.config import GeneralConfig, GeneralUserConfig
@@ -44,59 +44,6 @@ from app.utils.constants import UTC4
 from .tools import execute_script_task
 
 logger = get_logger("通用脚本自动代理")
-
-
-def _build_hook_logger(module: ModuleType, raw_path: str):
-    """为 Hook 模块构建专属 logger（优先使用 HOOK_META.name）。"""
-    hook_name: str | None = None
-    meta = getattr(module, "HOOK_META", None)
-    if isinstance(meta, dict):
-        name = meta.get("name")
-        if isinstance(name, str) and name.strip():
-            hook_name = name.strip()
-
-    if not hook_name:
-        stem = Path(raw_path).stem.strip()
-        hook_name = stem or getattr(module, "__name__", "unknown_hook")
-
-    return get_logger(f"Hook:{hook_name}")
-
-
-def _load_hook_module_from_path(file_path: str) -> tuple[ModuleType | None, str | None]:
-    """从 .py 文件路径动态加载模块。
-
-    返回 (module, warning)。失败时 module 为 None，warning 为说明文本。
-    """
-    p = Path(file_path)
-    if not p.exists() or not p.is_file():
-        return None, "文件不存在或不是文件"
-    if p.suffix.lower() != ".py":
-        return None, "仅支持 .py 文件"
-
-    try:
-        resolved = p.resolve()
-    except Exception:
-        resolved = p
-
-    # 生成稳定且尽量不冲突的模块名（避免重复 import 相互覆盖）
-    digest = sha1(str(resolved).encode("utf-8", errors="ignore")).hexdigest()[:12]
-    module_name = f"auto_mas_hook_{p.stem}_{digest}"
-
-    # 每次任务运行都重新加载一份，避免被上次运行的模块状态污染。
-    if module_name in sys.modules:
-        sys.modules.pop(module_name, None)
-
-    spec = importlib.util.spec_from_file_location(module_name, str(resolved))
-    if spec is None or spec.loader is None:
-        return None, "无法创建模块加载器"
-
-    module = importlib.util.module_from_spec(spec)
-    try:
-        spec.loader.exec_module(module)
-    except Exception as e:
-        return None, f"导入失败: {type(e).__name__}: {e}"
-
-    return module, None
 
 
 class AutoProxyTask(TaskExecuteBase):
@@ -152,66 +99,11 @@ class AutoProxyTask(TaskExecuteBase):
 
         if not hook_paths:
             return
-
-        for raw_path in hook_paths:
-            module, warn = _load_hook_module_from_path(raw_path)
-            if warn:
-                msg = f"Hook 加载警告 [{raw_path}]: {warn}"
-                self._hook_warnings.append(msg)
-                logger.warning(msg)
-                continue
-
-            assert module is not None
-            hook_logger = _build_hook_logger(module, raw_path)
-            # 向 Hook 模块注入可复用 logger，便于 Hook 内统一记录日志而非 print。
-            setattr(module, "HOOK_LOGGER", hook_logger)
-
-            set_logger = getattr(module, "set_logger", None)
-            if callable(set_logger):
-                try:
-                    set_logger(hook_logger)
-                except Exception as e:
-                    msg = f"Hook 日志注入警告 [{raw_path}]: {type(e).__name__}: {e}"
-                    self._hook_warnings.append(msg)
-                    logger.warning(msg)
-
-            register = getattr(module, "register", None)
-            if not callable(register):
-                msg = f"Hook 加载警告 [{raw_path}]: 未找到可调用的 register(scope, target_cls)"
-                self._hook_warnings.append(msg)
-                logger.warning(msg)
-                continue
-
-            try:
-                # 兼容两种签名：
-                # - register(scope, target_cls)
-                # - register(scope, target_cls, hook_logger)
-                sig = inspect.signature(register)
-                params = list(sig.parameters.values())
-                positional_count = len(
-                    [
-                        p
-                        for p in params
-                        if p.kind
-                        in (
-                            inspect.Parameter.POSITIONAL_ONLY,
-                            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-                        )
-                    ]
-                )
-                has_varargs = any(
-                    p.kind == inspect.Parameter.VAR_POSITIONAL for p in params
-                )
-
-                if has_varargs or positional_count >= 3:
-                    register(self._hook_scope, self.__class__, hook_logger)
-                else:
-                    register(self._hook_scope, self.__class__)
-            except Exception as e:
-                msg = f"Hook 注册警告 [{raw_path}]: {type(e).__name__}: {e}"
-                self._hook_warnings.append(msg)
-                logger.warning(msg)
-                continue
+        self._hook_warnings = load_hook_module(
+            hook_paths,
+            self._hook_scope,
+            self.__class__,
+        )
 
     async def check(self) -> str:
         if self.script_config.get(
@@ -715,13 +607,6 @@ class AutoProxyTask(TaskExecuteBase):
             self.cur_user_item.status = "异常"
 
     async def on_crash(self, e: Exception):
-        # 崩溃时也要确保退出 HookScope
-        if self._hook_scope is not None:
-            try:
-                await self._hook_scope.__aexit__(type(e), e, e.__traceback__)
-            finally:
-                self._hook_scope = None
-
         self.cur_user_item.status = "异常"
         logger.exception(f"自动代理任务出现异常: {e}")
         await Config.send_websocket_message(
